+++ internal/database/database.go
@@ -12,36 +12,8 @@ import (
 	"gorm.io/gorm/logger"
 )
 
-const defaultSystemPrompt = `{{if .CurrentTime}}current_time="{{.CurrentTime}}"
-{{end}}{{if .ChatTitle}}chat_title="{{.ChatTitle}}"
-{{end}}{{if .ChatType}}chat_type="{{.ChatType}}"
-{{end}}
-# Begin System Directives
-
-Your name is Tellama.
-You are an AI chatbot built by K4YT3X for Telegram group chats.
-Your task is to help users by providing information and answering questions.
-You must not engage in any harmful, illegal, or unethical conversations.
-You must be polite, respectful, and helpful to all users.
-You must obey laws, morals, and ethics.
-You should respond in plain text.
-
-# End System Directives`
-
 type DatabaseManager struct {
 	db *gorm.DB
-}
-
-type ChatMessage struct {
-	Timestamp time.Time
-	ChatID    int64
-	ChatTitle string
-	Role      string
-	UserID    int64
-	Username  string
-	FirstName string
-	LastName  string
-	Content   string
 }
 
 type AllowedChat struct {
@@ -50,9 +22,13 @@ type AllowedChat struct {
 	ChatTitle string `gorm:"unique"`
 }
 
-type SystemPrompt struct {
-	ID           uint   `gorm:"primaryKey;autoIncrement"`
-	ChatID       *int64 `gorm:"unique"`
+type ChatOverride struct {
+	ID           uint  `gorm:"primaryKey;autoIncrement"`
+	ChatID       int64 `gorm:"unique"`
+	ChatTitle    string
+	OllamaHost   string
+	Model        string
+	Options      string
 	SystemPrompt string
 }
 
@@ -66,19 +42,20 @@ type Message struct {
 	Username  string
 	FirstName string
 	LastName  string
-	Message   string
+	Content   string
 }
 
 type GenerationRequest struct {
-	ID        uint      `gorm:"primaryKey;autoIncrement"`
-	Timestamp time.Time `gorm:"autoCreateTime"`
-	ChatID    int64     `gorm:"index"`
-	ChatTitle string
-	UserID    int64
-	Username  string
-	Model     string
-	Options   string
-	Prompt    string
+	ID         uint      `gorm:"primaryKey;autoIncrement"`
+	Timestamp  time.Time `gorm:"autoCreateTime"`
+	ChatID     int64     `gorm:"index"`
+	ChatTitle  string
+	UserID     int64
+	Username   string
+	Model      string
+	Options    string
+	Prompt     string
+	OllamaHost string
 }
 
 func NewDatabaseManager(dbPath string) (*DatabaseManager, error) {
@@ -89,7 +66,7 @@ func NewDatabaseManager(dbPath string) (*DatabaseManager, error) {
 		return nil, fmt.Errorf("failed to open database: %w", err)
 	}
 
-	err = db.AutoMigrate(&AllowedChat{}, &SystemPrompt{}, &Message{}, &GenerationRequest{})
+	err = db.AutoMigrate(&AllowedChat{}, &ChatOverride{}, &Message{}, &GenerationRequest{})
 	if err != nil {
 		return nil, fmt.Errorf("failed to migrate tables: %w", err)
 	}
@@ -103,43 +80,99 @@ func (dm *DatabaseManager) IsChatAllowed(chatID int64) bool {
 	return !errors.Is(result.Error, gorm.ErrRecordNotFound)
 }
 
-func (dm *DatabaseManager) getDefaultSystemPrompt() string {
-	var systemPrompt SystemPrompt
-	result := dm.db.Where("chat_id IS NULL").First(&systemPrompt)
+func (dm *DatabaseManager) GetGlobalChatOverride() (ChatOverride, error) {
+	var chatOverride ChatOverride
+	result := dm.db.Where("chat_id IS NULL").First(&chatOverride)
+	if result.Error != nil && !errors.Is(result.Error, gorm.ErrRecordNotFound) {
+		return chatOverride, result.Error
+	}
+	return chatOverride, nil
+}
+
+func (dm *DatabaseManager) GetChatOverride(chatID int64) (ChatOverride, error) {
+	// Get the default chat override
+	globalChatOverride, err := dm.GetGlobalChatOverride()
+	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
+		return globalChatOverride, err
+	}
+
+	// Get the chat-specific override
+	var chatOverride ChatOverride
+	result := dm.db.Where("chat_id = ?", chatID).First(&chatOverride)
 	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-		return defaultSystemPrompt
+		return globalChatOverride, nil
+	} else if result.Error != nil {
+		return chatOverride, result.Error
 	}
-	if result.Error != nil {
-		return defaultSystemPrompt
+
+	// Merge non-empty fields from chatOverride into globalChatOverride
+	globalChatOverride.ChatID = chatOverride.ChatID
+	if chatOverride.ChatTitle != "" {
+		globalChatOverride.ChatTitle = chatOverride.ChatTitle
 	}
-	return systemPrompt.SystemPrompt
+	if chatOverride.OllamaHost != "" {
+		globalChatOverride.OllamaHost = chatOverride.OllamaHost
+	}
+	if chatOverride.Model != "" {
+		globalChatOverride.Model = chatOverride.Model
+	}
+	if chatOverride.Options != "" {
+		globalChatOverride.Options = chatOverride.Options
+	}
+	if chatOverride.SystemPrompt != "" {
+		globalChatOverride.SystemPrompt = chatOverride.SystemPrompt
+	}
+
+	return globalChatOverride, nil
 }
 
-func (dm *DatabaseManager) GetSystemPromptForGroup(chatID int64) string {
-	var systemPrompt SystemPrompt
-	result := dm.db.Where("chat_id = ?", chatID).First(&systemPrompt)
-	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-		return dm.getDefaultSystemPrompt()
+func (dm *DatabaseManager) SetChatOverride(
+	chatID int64,
+	chatTitle string,
+	ollamaHost string,
+	model string,
+	options string,
+	systemPrompt string,
+) error {
+	chatOverride := ChatOverride{
+		ChatID: chatID,
 	}
-	if result.Error != nil {
-		return defaultSystemPrompt
+
+	// Prepare the map of columns to update only if the field is non-empty
+	updates := map[string]interface{}{}
+
+	if chatTitle != "" {
+		chatOverride.ChatTitle = chatTitle
+		updates["chat_title"] = chatTitle
 	}
-	return systemPrompt.SystemPrompt
+	if ollamaHost != "" {
+		chatOverride.OllamaHost = ollamaHost
+		updates["ollama_host"] = ollamaHost
+	}
+	if model != "" {
+		chatOverride.Model = model
+		updates["model"] = model
+	}
+	if options != "" {
+		chatOverride.Options = options
+		updates["options"] = options
+	}
+	if systemPrompt != "" {
+		chatOverride.SystemPrompt = systemPrompt
+		updates["system_prompt"] = systemPrompt
+	}
+
+	// Update only the non-empty fields
+	return dm.db.Clauses(
+		clause.OnConflict{
+			Columns:   []clause.Column{{Name: "chat_id"}},
+			DoUpdates: clause.Assignments(updates),
+		},
+	).Create(&chatOverride).Error
 }
 
-func (dm *DatabaseManager) SetSystemPromptForGroup(chatID int64, systemPrompt string) error {
-	sp := SystemPrompt{
-		ChatID:       &chatID,
-		SystemPrompt: systemPrompt,
-	}
-	return dm.db.Clauses(clause.OnConflict{
-		Columns:   []clause.Column{{Name: "chat_id"}},
-		DoUpdates: clause.AssignmentColumns([]string{"system_prompt"}),
-	}).Create(&sp).Error
-}
-
-func (dm *DatabaseManager) DeleteSystemPromptForGroup(chatID int64) error {
-	return dm.db.Where("chat_id = ?", chatID).Delete(&SystemPrompt{}).Error
+func (dm *DatabaseManager) DeleteChatOverride(chatID int64) error {
+	return dm.db.Where("chat_id = ?", chatID).Delete(&ChatOverride{}).Error
 }
 
 func (dm *DatabaseManager) StoreMessage(
@@ -160,11 +193,11 @@ func (dm *DatabaseManager) StoreMessage(
 		Username:  username,
 		FirstName: firstName,
 		LastName:  lastName,
-		Message:   messageText,
+		Content:   messageText,
 	}).Error
 }
 
-func (dm *DatabaseManager) GetMessages(chatID int64, limit int) ([]ChatMessage, error) {
+func (dm *DatabaseManager) GetMessages(chatID int64, limit int) ([]Message, error) {
 	var messages []Message
 	result := dm.db.Where("chat_id = ?", chatID).
 		Order("id DESC").
@@ -174,9 +207,9 @@ func (dm *DatabaseManager) GetMessages(chatID int64, limit int) ([]ChatMessage, 
 		return nil, result.Error
 	}
 
-	history := make([]ChatMessage, len(messages))
+	history := make([]Message, len(messages))
 	for i, m := range messages {
-		history[i] = ChatMessage{
+		history[i] = Message{
 			Timestamp: m.Timestamp,
 			ChatID:    m.ChatID,
 			ChatTitle: m.ChatTitle,
@@ -185,7 +218,7 @@ func (dm *DatabaseManager) GetMessages(chatID int64, limit int) ([]ChatMessage, 
 			Username:  m.Username,
 			FirstName: m.FirstName,
 			LastName:  m.LastName,
-			Content:   m.Message,
+			Content:   m.Content,
 		}
 	}
 
@@ -205,14 +238,16 @@ func (dm *DatabaseManager) StoreGenerationRequest(
 	model string,
 	options string,
 	prompt string,
+	ollamaHost string,
 ) error {
 	return dm.db.Create(&GenerationRequest{
-		ChatID:    chatID,
-		ChatTitle: chatTitle,
-		UserID:    userID,
-		Username:  username,
-		Model:     model,
-		Options:   options,
-		Prompt:    prompt,
+		ChatID:     chatID,
+		ChatTitle:  chatTitle,
+		UserID:     userID,
+		Username:   username,
+		Model:      model,
+		Options:    options,
+		Prompt:     prompt,
+		OllamaHost: ollamaHost,
 	}).Error
 }